updated_with_event_in_storage
