--- src/utils/localStorageInterceptor.ts
@@ -1,10 +1,0 @@
-import { fillLocalStorage } from '@/utils/storage';
-
-// TODO delete later
-// fillLocalStorage(96);
-
-// const originalSetItem = localStorage.setItem;
-// localStorage.setItem = function (key: string, value: string) {
-//   window.dispatchEvent(new Event('localStorageUpdated'));
-//   originalSetItem.call(this, key, value);
-// };
